Catalogo-Sagar/requirements.txt
flask==3.1.2
python-dotenv==1.2.1
supabase==2.27.2
gunicorn>=21.2.0